Separate args related function into args.go
